Add tests for Spring validation.Errors
