+++ java/ql/test/library-tests/frameworks/spring/validation/options
@@ -1,0 +1,1 @@
+//semmle-extractor-options: --javac-args -cp ${testdir}/../../../../stubs/springframework-5.3.8
